Module developer_hub::guides

In-depth guides about the most common components of the Polkadot SDK. They are slightly more high level and broad than reference docs.

Polkadot Developer Hub Guides

This crate contains a collection of guides that are foundational to the developers of Polkadot SDK. They common user-journeys that are traversed in the Polkadot ecosystem.

Modules

changing_consensus

How to change the consensus engine of both the node and the runtime.

cumulus_enabled_parachain

How to enhance a given runtime and node to be cumulus-enabled, run it as a parachain and connect it to a relay-chain.

xcm_enabled_parachain

How to make a given runtime XCM-enabled, capable of sending messages (Transact) between itself and the relay chain to which it is connected.

your_first_node

Running the given runtime with a node. No specific consensus mechanism is used at this stage.

your_first_pallet

Write your first simple pallet, learning the most most basic features of FRAME along the way.

your_first_runtime

Writing your first real runtime, and successfully compiling it to WASM.