Module developer_hub::polkadot_sdk

An introduction to the Polkadot SDK. Read this module to learn about the structure of the SDK, the tools that are provided as a part of it, and to gain a high level understanding of each.

Polkadot SDK

The Polkadot SDK repository provides all the resources needed to start building on the Polkadot network, a multi-chain blockchain platform that enables different blockchains to interoperate and share information in a secure and scalable way.

Getting Started

The primary way to get started with the Polkadot SDK is to start writing a Substrate-based runtime using FRAME. See:

polkadot, to understand what is Polkadot as a development platform.

substrate, for an overview of what Substrate is.

Jump right into frame to learn about how to write a FRAME-based blockchain runtime.

Continue with the developer_hub’s “getting started”.

Structure

This repository is a nested workspace, containing the following major software artifacts:

Substrate

substrate is the base blockchain framework used to power the Polkadot SDK. It is a full toolkit to create sovereign blockchains, including but not limited to those who connect to Polkadot as parachains.

FRAME

frame is the framework used to create Substrate-based runtimes. Learn more

Cumulus

cumulus transforms FRAME-based runtimes into Polkadot-compatible parachain runtimes, and substrate-based client into Polkadot-compatible collators.

Polkadot

Recall from Substrate’s architecture that any substrate-based chain is composed of two parts: A client, and a runtime.

polkadot is an implementation of a Polkadot client in Rust, by @paritytech. The Polkadot runtimes are located under the fellowship/runtime repository.

polkadot contains useful links to further learn about Polkadot, but is in general not part of the SDK, as it is rarely used by developers who wish to build on top of Polkadot.

The following diagram summarizes how components of the Polkadot-SDK work together:

flowchart LR
    subgraph Parachain[A Polkadot Parachain]
        ParachainCollator[Parachain Collator]
        ParachainRuntime[Parachain Runtime]
    end

    subgraph Polkadot[The Polkadot Relay Chain]
        PolkadotClinet[Polkadot Clinet]
        PolkadotRuntime[Polkadot Runtime]
    end

    subgraph SubstrateChain[A Substrate-based blockchain]
        Clinet
        Runtime
    end

    FRAME1[FRAME] -.-> Runtime
    FRAME2[FRAME] -.-> PolkadotRuntime
    FRAME3[FRAME] -.-> ParachainRuntime

    Substrate1[Substrate Client Libraries] -.-> Clinet
    Substrate2[Substrate Client Libraries] -.-> PolkadotClinet
    Substrate3[Substrate Client Libraries] -.-> ParachainCollator

    Cumulus1[Cumulus Runtime Pallets/Libraries] -.-> ParachainRuntime
    Cumulus2[Cumulus Client Libraries] -.-> ParachainCollator

A Substrate-based chain is a blockchain composed of a “Runtime” and a “Client”. As noted above, the “Runtime” is the application logic of the blockchain, and the “Client” is everything else. See reference_docs::wasm_meta_protocol for an in-depth explanation of this. The former is built with frame, and the latter is built with Substrate client libraries.

Polkadot is itself a Substrate-based chain, composed of the the exact same two components. The Polkadot client code is in polkadot, and the Polkadot runtimes are controlled by the Polkadot Fellowship.

A parachain is a “special” Substrate based chain, whereby both the client and the runtime components have became “Polkadot-aware” using Cumulus.

History

Substrate, Polkadot and Cumulus used to each have their own repository, each of which is now archived. For historical context about how they merged into this mono-repo, see:

https://polkadot-public.notion.site/Polkadot-SDK-FAQ-fbc4cecc2c46443fb37b9eeec2f0d85f

https://forum.polkadot.network/t/psa-parity-is-currently-working-on-merging-the-polkadot-stack-repositories-into-one-single-repository/2883

Notable Upstream Crates:

parity-scale-codec

parity-db

trie

parity-common

Modules

cumulus

Lean about Cumulus, the framework that transforms substrate-based chains into polkadot-enabled parachains.

frame_runtime

Learn about FRAME, the framework used to build Substrate runtimes.

polkadot

Learn about Polkadot as a platform.

smart_contracts

Learn about different ways through which smart contracts can be utilized on top of Substrate, and in the Polkadot ecosystem. Under Construction

substrate

Learn about Substrate, the main blockchain framework used in the Polkadot ecosystem.

templates

Index of all the templates that can act as an initial scaffold for a new project.

xcm

Learn about XCM, the de-facto communication language between different consensus systems.