Module developer_hub::reference_docs

Reference documents covering in-depth topics across the Polkadot SDK. It is suggested to read these on-demand, while you are going through the tutorial or other content.

Polkadot SDK Reference Docs.

This is the entry point for all reference documents that enhance one’s learning experience in the Polkadot SDK.

What is a “reference document”?

First, see why we use rust-docs for everything and our documentation principles. We acknowledge that as much of the crucial information should be embedded in the low level rust-docs. Then, high level scenarios should be covered in crate::tutorial. Finally, we acknowledge that there is a category of information that is:

crucial to know.

is too high level to be in the rust-doc of any one type, trait or fn.

is too low level to be encompassed in a crate::tutorial.

We can this class of documents “reference documents”. Our goal should be to minimize the number of “reference” docs, as they incur maintenance burden.

Ownership

Every page must have an owner or a list of owners, who are responsible for maintaining the page.

Modules

blockchain_state_machines

Learn about the way Substrate and FRAME view their blockchains as state machines.

extrinsic_encoding

Learn about how extrinsics are encoded to be transmitted to a node and stored in blocks.

fee_less_runtime

Learn about how to make a pallet/runtime that is fee-less and instead uses another mechanism to control usage and sybil attacks.

frame_composite_enums

Learn about composite enums in FRAME-based runtimes, such as “RuntimeEvent” and “RuntimeCall”.

glossary

The glossary.

metadata

Learn about metadata, the main means through which an upgrade-able runtime communicate its properties to the outside world.

origin_account_abstraction

Learn about “Origin” A topic in FRAME that enables complex account abstractions to be built.

runtime_vs_smart_contract

Learn about the differences between smart contracts and a FRAME-based runtime. They are both “code stored onchain”, but how do they differ?

safe_defensive_programming

Learn about how to write safe and defensive code in your FRAME runtime.

signed_extensions

Learn about the signed extensions that form a part of extrinsics. Signed extensions are, briefly, a means for different chains to extend the “basic” extrinsic format with custom data that can be checked by the runtime.

trait_based_programming

Learn how Substrate and FRAME use traits and associated types to make modules generic in a type-safe manner.

wasm_meta_protocol

Learn about the WASM meta-protocol of all substrate-based chains.